Finds, keeps track and recognizes signs
Works. Motor moves are commented;  need (a lot of) calibration of the Dagu dc motors (weak point of the bot)
The script is a single run for detecting, simulating move towards the sign (just 10 new readings, with adjustment of the neck) and finding the right comparison. (The while-loop for != STOP to be inserted rather easely ).
The script is (over) commented, so, I hope, self-explaining
Next sprint: calibrating the moves (pffff....), tweaking performance, construct a local script (run without the use of websockets)